Code of Chatbot Poject/requirements.txt
flask
redis
line-bot-sdk
requests